Revert "Feature/today refactor"
