+++ src/components/layoutPieces/SectionTitle/styles.css
@@ -1,9 +1,9 @@
-@import '../../../themes/index.css';
+@import "../../../themes/index.css";
 
 .sectionTitleText {
-  color: var(--darkBlue);
-  font-family: var(--subtitleFont);
-  font-size: var(--subtitleSize);
-  font-weight: var(--subtitleWeight);
-  margin: var(--elementPadding) 0;
-}
+    color: var(--darkBlue);
+    font-family: var(--subtitleFont);
+    font-size: var(--subtitleSize);
+    font-weight: var(--subtitleWeight);
+    margin: var(--elementPadding) 0;
+  }
