+++ src/components/layoutPieces/index.js
@@ -1,2 +1,2 @@
-export { default as MainTitle } from './MainTitle';
-export { default as SectionTitle } from './SectionTitle';
+export { default as MainTitle } from './MainTitle/';
+export { default as SectionTitle } from './SectionTitle/';
