+++ src/pages/todayStyles.css
@@ -1,4 +1,4 @@
-@import '../styles.css';
+@import './styles.css';
 
 .leftView {
   height: 100%;
