--- src/utils/index.js
@@ -1,3 +1,0 @@
-export * from './formatters';
-export { default as NOMES_PROMOTORIAS } from './nomesPromotorias';
-export { default as dataStateWrapper } from './dataStateWrapper';
